+++ soulBank/soulBank/src/main/java/codingDreams/model/PessoaFisica.java
@@ -5,7 +5,7 @@ public class PessoaFisica extends Cliente{
     private String rg;
     private String nome;
 
-    public PessoaFisica(int idCliente, String email, String telefone, Endereco endereco, ContaBancaria contaBancaria, String cpf, String rg, String nome) {
+    public PessoaFisica(Long idCliente, String email, String telefone, Endereco endereco, ContaBancaria contaBancaria, String cpf, String rg, String nome) {
         super(idCliente, email, telefone, endereco, contaBancaria);
         this.cpf = cpf;
         this.rg = rg;
